+++ sipes/src/main/java/apap/tugas/sipes/repository/TipeDb.java
@@ -9,7 +9,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TipeDb extends JpaRepository<TipeModel,BigInteger>{
+public interface TipeDb extends JpaRepository<TipeModel,Long>{
     Optional<TipeModel> findById(Long id);
-
 }
